Move common tests back
